requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.3.0
boto3==1.38.13
botocore==1.38.13
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dictdiffer==0.9.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iso8601==2.1.0
jmespath==1.0.1
mypy-protobuf==3.6.0
protobuf==5.29.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tortoise-orm==0.24.2
types-protobuf==5.29.1.20250403
types-pyasn1==0.6.0.20250208
types-python-jose==3.4.0.20250224
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
